+++ dashboard/src/store/modules/restqa/state.js
@@ -1,0 +1,7 @@
+const state = {
+  config: null,
+  loadingConfig: false,
+  selectedEnv: null
+}
+
+export default state