+++ dashboard/src/store/store.js
@@ -5,6 +5,7 @@ import state from './state'
 import getters from './getters'
 import mutations from './mutations'
 import actions from './actions'
+import restqa from './modules/restqa'
 
 Vue.use(Vuex)
 
@@ -15,6 +16,7 @@ export default new Vuex.Store({
   state,
   actions,
   modules: {
+    restqa
   },
   strict: process.env.NODE_ENV !== 'production'
 })