+++ src/http-server/index.test.js
@@ -1,11 +1,15 @@
+const nock = require('nock')
 const request = require('supertest')
 const path = require('path')
 const fs = require('fs')
 const os = require('os')
 const rimraf = require('rimraf')
 const { EventEmitter } = require('events')
+const YAML = require('yaml')
 
 let filename
+
+const jestqa = new JestQA(__filename, false)
 
 afterEach(() => {
   if (filename && fs.existsSync(filename)) {
@@ -26,11 +30,15 @@ afterAll(() => {
   if (buggedReportforlder && fs.existsSync(buggedReportforlder)) {
     rimraf.sync(buggedReportforlder)
   }
+  nock.cleanAll()
+  nock.enableNetConnect()
 })
 
 jest.mock('../utils/logger', () => {
   return {
-    info: jest.fn()
+    info: jest.fn(),
+    log: jest.fn(),
+    success: jest.fn()
   }
 })
 
@@ -83,9 +91,56 @@ describe('#dashboard > Server', () => {
     })
   })
 
+  describe('/config', () => {
+    test('throw error if server is running on "NO CONFIG" mode', async () => {
+      const config = false
+      const response = await request(server(config)).get('/config')
+      expect(response.status).toBe(403)
+      expect(response.body.message).toBe('Please initiate your RestQA project before using this endpoint.')
+    })
+
+    test('Return the configuration', async () => {
+      const content = `
+---
+
+version: 0.0.1
+metadata:
+  code: API
+  name: My test API
+  description: The decription of the test api
+environments:
+  - name: local
+    default: true
+    plugins:
+      - name: 'restqa/restqapi'
+        config:
+          url: http://localhost:3000
+    outputs:
+      - type: file
+        enabled: true
+        config:
+          path: 'my-report.json'
+      `
+      filename = path.resolve(os.tmpdir(), '.restqa.yml')
+      fs.writeFileSync(filename, content)
+
+      const response = await request(server(filename))
+        .get('/config')
+      expect(response.status).toBe(200)
+      expect(response.body).toEqual(YAML.parse(content))
+    })
+  })
+
   describe('/api/steps', () => {
+    test('throw error if server is running on "NO CONFIG" mode', async () => {
+      const config = false
+      const response = await request(server(config)).get('/api/restqa/steps?keyword=cool')
+      expect(response.status).toBe(403)
+      expect(response.body.message).toBe('Please initiate your RestQA project before using this endpoint.')
+    })
+
     test('throw error if the keyword is not incorrect', async () => {
-      const config = {}
+      const config = '.restqa.yml'
       const response = await request(server(config)).get('/api/restqa/steps?keyword=cool')
       expect(response.status).toBe(406)
       expect(response.body.message).toBe('"cool" is not a valid argument. Available: given | when | then')
@@ -161,6 +216,110 @@ environments:
     })
   })
 
+  describe('/api/initialize', () => {
+    test('throw error if name is not defined', async () => {
+      const folder = jestqa.getTmpFolder()
+      const config = {}
+      const response = await request(server(config))
+        .post('/api/restqa/initialize')
+        .send({
+          folder
+        })
+      expect(response.status).toBe(406)
+      expect(response.body.message).toBe('Please share a project name.')
+    })
+
+    test('throw error if description is not defined', async () => {
+      const folder = jestqa.getTmpFolder()
+      const config = {}
+      const response = await request(server(config))
+        .post('/api/restqa/initialize')
+        .send({
+          name: 'Backend api',
+          folder
+        })
+      expect(response.status).toBe(406)
+      expect(response.body.message).toBe('Please share a project description.')
+    })
+
+    test('throw error if url is not defined', async () => {
+      const folder = jestqa.getTmpFolder()
+      const config = {}
+      const response = await request(server(config))
+        .post('/api/restqa/initialize')
+        .send({
+          name: 'Backend api',
+          description: 'All the API used by the different frontends',
+          folder
+        })
+      expect(response.status).toBe(406)
+      expect(response.body.message).toBe('Please share a project url.')
+    })
+
+    test('throw error if env is not defined', async () => {
+      const folder = jestqa.getTmpFolder()
+      const config = {}
+      const response = await request(server(config))
+        .post('/api/restqa/initialize')
+        .send({
+          name: 'Backend api',
+          description: 'All the API used by the different frontends',
+          url: 'https://api.example.com',
+          folder
+        })
+      expect(response.status).toBe(406)
+      expect(response.body.message).toBe('Please share a project url environment.')
+    })
+
+    test('throw error if ci tool is invalid', async () => {
+      const folder = jestqa.getTmpFolder()
+      const config = {}
+      const response = await request(server(config))
+        .post('/api/restqa/initialize')
+        .send({
+          name: 'Backend api',
+          description: 'All the API used by the different frontends',
+          url: 'https://api.example.com',
+          ci: 'gocd',
+          env: 'uat',
+          folder
+        })
+      expect(response.status).toBe(406)
+      expect(response.body.message).toBe('The continous integration "gocd" is not supported by RestQa')
+    })
+
+    test('Create a new restqa project', async () => {
+      nock('https://restqa.io')
+        .get('/welcome.json')
+        .reply(200, {
+          foo: 'bar'
+        })
+
+      const folder = jestqa.getTmpFolder()
+      const config = {}
+      const srv = server(config)
+      const response = await request(srv)
+        .post('/api/restqa/initialize')
+        .send({
+          name: 'Backend api',
+          description: 'All the API used by the different frontends',
+          url: 'https://api.example.com',
+          env: 'uat',
+          ci: 'gitlab-ci',
+          folder
+        })
+      expect(response.status).toBe(200)
+      expect(response.body.configuration).toBe(path.join(folder, '.restqa.yml'))
+      expect(response.body.folder).toEqual(folder)
+      expect(fs.existsSync(path.join(folder, '.restqa.yml'))).toBe(true)
+      expect(fs.existsSync(path.join(folder, '.gitlab-ci.yml'))).toBe(true)
+      expect(fs.existsSync(path.join(folder, 'tests', 'integration', 'welcome-restqa.feature'))).toBe(true)
+
+      const responseConfig = await request(srv).get('/config')
+      expect(responseConfig.status).toBe(200)
+    })
+  })
+
   describe('/api/generate', () => {
     test('throw error if the command is not a curl command', async () => {
       const config = {}
@@ -172,6 +331,15 @@ environments:
     })
 
     test('Generate the curl command', async () => {
+      nock('https://jsonplaceholder.typicode.com')
+        .get('/todos/1')
+        .reply(200, {
+          userId: 1,
+          id: 1,
+          title: 'delectus aut autem',
+          completed: false
+        })
+
       filename = path.resolve(os.tmpdir(), '.restqa.yml')
 
       const response = await request(server(filename))
@@ -200,6 +368,13 @@ Then I should receive a response with the status 200
   })
 
   describe('/api/install', () => {
+    test('throw error if server is running on "NO CONFIG" mode', async () => {
+      const config = false
+      const response = await request(server(config)).post('/api/restqa/install')
+      expect(response.status).toBe(403)
+      expect(response.body.message).toBe('Please initiate your RestQA project before using this endpoint.')
+    })
+
     test('throw error if the integration to install doesn\'t exist', async () => {
       const config = './restqa.yml'
       const options = {
@@ -334,6 +509,13 @@ environments:
   })
 
   describe('/api/run', () => {
+    test('throw error if server is running on "NO CONFIG" mode', async () => {
+      const config = false
+      const response = await request(server(config)).post('/api/restqa/run')
+      expect(response.status).toBe(403)
+      expect(response.body.message).toBe('Please initiate your RestQA project before using this endpoint.')
+    })
+
     test('throw error if the configuration file  doesn\'t exist', async () => {
       const config = './.restqa.yml'
       const options = {