+++ src/index.test.js
@@ -69,6 +69,41 @@ describe('# Index - Generate', () => {
   })
 })
 
+describe('# Index - Initialize', () => {
+  test('Throw an error if ci is not valid', () => {
+    const options = {
+      url: 'http://test.com',
+      env: 'test',
+      description: 'my description',
+      folder: jestqa.getTmpFolder()
+    }
+    const { Initialize } = require('./index')
+    return expect(Initialize(options)).rejects.toThrow('Please share a project name.')
+  })
+
+  test('Initiate a new project', async () => {
+    jestqa.getLoggerMock()
+    const mockGenerate = jest.fn().mockResolvedValue('result')
+    jest.mock('./cli/generate', () => {
+      return mockGenerate
+    })
+    const folder = jestqa.getTmpFolder()
+    const options = {
+      name: 'sample',
+      url: 'http://test.com',
+      env: 'test',
+      description: 'my description',
+      ci: 'gitlab-ci',
+      folder
+    }
+    const { Initialize } = require('./index')
+    const result = await Initialize(options)
+    expect(result).toEqual(path.join(folder, '.restqa.yml'))
+    expect(fs.existsSync(path.join(folder, '.gitlab-ci.yml'))).toBe(true)
+    expect(fs.existsSync(path.join(folder, 'tests', 'integration', 'welcome-restqa.feature'))).toBe(true)
+  })
+})
+
 describe('# Index - Install', () => {
   test('Get result from an addon  Install', () => {
     const mockInstall = {