+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Improve command line performance | [#95](https://github.com/restqa/restqa/pull/95)
 * Improve the installation user experience | [#94](https://github.com/restqa/restqa/pull/94)
 * [Enable Custom Webhook Alerting](https://docs.restqa.io/alerting/webhook) | [#99](https://github.com/restqa/restqa/pull/99)
+* Enable the project initialization from the dashboard | [#100](https://github.com/restqa/restqa/pull/100)
 
 ### Bugfix